Description
Nowadays, computer-based multiple-choice exams are used nearly
universally in the recruitment, examination, and other processes. Making
the setting for a computer-based test is very expensive because it requires
a lot of equipment and software. We are proposing a project that will use
computer vision to lessen the hardware setup required for computer-based
exams, such as a keyboard and mouse. The proposed project will first save
the question to a csv file before utilizing computer vision to show it on
screen. Secondly, using a webcam and computer vision, we will look for
hand gestures in the air that can be utilized to answer the questions that are
being displayed. Finally, the automatic grading mechanism in it will reveal
the grade. This method made it simple to answer questions and even made
the evaluation process simple. Importantly, because of very little
technology is needed, the exam won't be terminated in the middle due to a
hardware fault or malfunction